🏷️🔥 Increase exported types
- Exports more types from the top level indexes
- These are required to compile downstream typescript projects that use things in any depth, and especially if they re-export inferred types, e.g. returning the results of an `encrypt` method.
- While I'm here, I've cleaned up and simplified a few methods, and removed a couple of things that aren't used
