+++ lib/tdf3/index-web.ts
@@ -1,4 +1,23 @@
+import { FileClient } from './src/FileClient.js';
+import { Binary } from './src/binary.js';
+import { DecoratedReadableStream } from './src/client/DecoratedReadableStream.js';
+import { BrowserTdfStream } from './src/client/BrowserTdfSteam.js';
+import {
+  type DecryptParams,
+  DecryptParamsBuilder,
+  type EncryptParams,
+  EncryptParamsBuilder,
+} from './src/client/builders.js';
+import { type SessionKeys, type ClientConfig, createSessionKeys } from './src/client/index.js';
+import { type AnyTdfStream, registerModuleType } from './src/client/tdf-stream.js';
+import { type DecryptResult, type EncryptResult } from './src/crypto/declarations.js';
 import { TDF, Client, Errors } from './src/index.js';
+import {
+  type KeyInfo,
+  SplitKey,
+  type EncryptionInformation,
+} from './src/models/encryption-information.js';
+import { AuthProvider, AppIdAuthProvider, HttpRequest } from '../src/auth/auth.js';
 import {
   NanoTDFClient,
   NanoTDFDatasetClient,
@@ -6,26 +25,37 @@ import {
   version,
   clientType,
 } from '../src/index.js';
-import { FileClient } from './src/FileClient.js';
-import { AuthProvider, AppIdAuthProvider, HttpRequest } from '../src/auth/auth.js';
-import { registerModuleType } from './src/client/tdf-stream.js';
-import { BrowserTdfStream } from './src/client/BrowserTdfSteam.js';
 
 registerModuleType(BrowserTdfStream);
 
 window.TDF = TDF;
 
 export {
-  TDF,
-  Errors,
+  AnyTdfStream,
+  AppIdAuthProvider,
+  AuthProvider,
+  AuthProviders,
+  Binary,
   Client,
-  version,
-  clientType,
+  ClientConfig,
+  DecoratedReadableStream,
+  DecryptParams,
+  DecryptParamsBuilder,
+  DecryptResult,
+  EncryptionInformation,
+  EncryptParams,
+  EncryptParamsBuilder,
+  EncryptResult,
+  Errors,
+  FileClient,
+  HttpRequest,
+  KeyInfo,
   NanoTDFClient,
   NanoTDFDatasetClient,
-  AuthProviders,
-  FileClient,
-  AuthProvider,
-  AppIdAuthProvider,
-  HttpRequest,
+  SessionKeys,
+  SplitKey,
+  TDF,
+  clientType,
+  createSessionKeys,
+  version,
 };
