--- lib/tdf3/src/client/validation/index.ts
@@ -1,1 +1,0 @@
-export { AttributeValidator } from './validators.js';
