--- lib/tdf3/src/client/validation/validators.ts
@@ -1,18 +1,0 @@
-import { runAttributesValidation } from './validations.js';
-import { AttributeValidationError, IllegalArgumentError } from '../../errors.js';
-
-type Attribute = { attribute: string };
-
-const AttributeValidator = (attributes: Attribute[]) => {
-  try {
-    runAttributesValidation(attributes);
-  } catch (err) {
-    if (err instanceof AttributeValidationError) {
-      throw new IllegalArgumentError(err.message);
-    } else {
-      throw err;
-    }
-  }
-};
-
-export { AttributeValidator };
